+++ components/shared/custome/c-input.tsx
@@ -1,16 +1,12 @@
 "use client";
 
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+import CInputField from "./c-input-field";
 
 interface ICinput extends IFieldInput {
-  validators?: {
-    onChange?: (args: any) => void;
-    onChangeAsync: (args: any) => void;
-    onChangeAsyncDebounceMs: number;
-  };
   name: string;
-  form: any;
+  label?: string;
+  validators?: any;
+  form: any; // NOTE if u know the type of form, please let me know or just open pr
 }
 
 export default function CInput({
@@ -18,6 +14,10 @@ export default function CInput({
   validators,
   label,
   form,
+  field,
+  iconSvg,
+  iconImg,
+  iconSize,
   ...props
 }: ICinput) {
   return (
@@ -26,21 +26,18 @@ export default function CInput({
       validators={validators}
       children={(field: any) => {
         return (
-          <div className="relative mb-[25px]">
-            {label && <Label>{label}</Label>}
-            <Input
-              {...props}
-              name={field.name}
-              value={field.state.value as string}
-              onBlur={field.handleBlur}
-              onChange={(e) => field.handleChange(e.target.value)}
-            />
-            {field.state.meta.errors.length ? (
-              <em className="absolute left-0 bottom-[-20px] text-xs text-red-500">
-                {field.state.meta.errors?.map((err: any) => err.message)?.[0]}
-              </em>
-            ) : null}
-          </div>
+          <CInputField
+            label={label}
+            name={name}
+            value={field.state.value as string}
+            onBlur={field.handleBlur}
+            onChange={(e) => field.handleChange(e.target.value)}
+            iconSvg={iconSvg}
+            iconImg={iconImg}
+            iconSize={iconSize}
+            field={field}
+            {...props}
+          />
         );
       }}
     />